transaction_service/transaction_service/transactions/serializers.py
from rest_framework import serializers
from .models import Transaction

class TransactionSerializer(serializers.ModelSerializer):
    class Meta:
        model = Transaction
        fields = ["id", "department", "amount", "created_by", "created_at", "is_confirmed"]
